requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
asgiref==3.6.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
cachetools==5.3.0 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
cffi==1.15.1
charset-normalizer==3.0.1 ; python_full_version >= '3.6.0'
colorama==0.4.6 ; platform_system == 'Windows'
cryptography==39.0.1 ; python_version >= '3.6'
django==4.1.7
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-restframework==0.0.1
djangorestframework==3.14.0
frozenlist==1.3.3 ; python_version >= '3.7'
google-api-core==2.11.0 ; python_version >= '3.7'
google-api-python-client==2.78.0
google-auth==2.16.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0 ; python_version >= '3.7'
httplib2==0.21.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.4 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
openai==0.26.5
pipfile==0.0.2
protobuf==4.22.0 ; python_version >= '3.7'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9 ; python_version >= '3.1'
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
rsa==4.9 ; python_version >= '3.6'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.3 ; python_version >= '3.5'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tzdata==2022.7 ; sys_platform == 'win32'
uritemplate==4.1.1 ; python_version >= '3.6'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.8.2 ; python_version >= '3.7'
